Changed type name of `number` to `double`
This had to be done since the term number does not really represent the
specific semantic meaning of that type and could be misunderstood.

Additionally we extracted a super class for NumericalElements in the
BoLang VM, so they can be easier distinguished from other non-numerical
types such as Booleans, Arrays, Strings.